Fix balls falling from map and improve code
I am not certain if balls escaping is fully fixed but it seems that
making border simpler, with just 4 cubes for each side fixed the problem
Also changed balls to continous collision detection as advised in tut
but it doesn't seem to change anything.

Improved code redability by going through children with foreach and
swaping GetComponent for its template version.
Also in functions like Update, Start, etc code is in functions so its
easier to quickly read what is happening