Improve depth meter scaling
Depth meter extends correctly based on device height and
distance beetwen labels is correct but where label of current
depth is pointing difers beetwen devices. On 1280x720 it is
slighlty above miner, on Z Fold it is inside blocks.

This inconsistency is not crucial but it should be fixed if time is left